refactor(calculator): calculator에서 delimiter 책임 분리
