+++ README.md
@@ -52,3 +52,13 @@
 
    - 커스텀 구분자가 존재하는 경우 커스텀 구분자를 함께 구분자로 사용하고 아닌 경우 기본 구분자만을 사용한다.
    - 결정된 구분자를 사용하여 분리된 문자열을 반환한다.
+
+### Delimiter
+
+구현 중 Calculator가 이름에 맞지 않는 책임을 많이 담당하고 있다는 것을 알게되었습니다.
+
+실제로 Calculator는 계산기라는 이름과 맞지 않게 구분자와 관련된 많은 책임을 담당하고 있습니다.
+
+자연스럽게 Calculator에는 구분자와 관련된 private method가 많아지게 되었고 테스트가 필요하지만 할 수 없는 class가 만들어지게 되었습니다.
+
+또한 구분자와 관련된 메서드들은 입력값을 검증하는데도 필요하기 때문에 따로 분리할 필요성을 느끼게 되었습니다.
